+++ yfinance/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-# Yahoo! Finance market data downloader (+fix for Pandas Datareader)
+# yfinance - market data downloader
 # https://github.com/ranaroussi/yfinance
 #
 # Copyright 2017-2019 Ran Aroussi
@@ -19,12 +19,13 @@
 # limitations under the License.
 #
 
-__version__ = "0.1.54"
-__author__ = "Ran Aroussi"
-
+from . import version
 from .ticker import Ticker
 from .tickers import Tickers
 from .multi import download
+
+__version__ = version.version
+__author__ = "Ran Aroussi"
 
 
 def pdr_override():