+++ yfinance/base.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-# Yahoo! Finance market data downloader (+fix for Pandas Datareader)
+# yfinance - market data downloader
 # https://github.com/ranaroussi/yfinance
 #
 # Copyright 2017-2019 Ran Aroussi
@@ -26,6 +26,7 @@ import datetime as _datetime
 import requests as _requests
 import pandas as _pd
 import numpy as _np
+import re as _re
 
 try:
     from urllib.parse import quote as urlencode
@@ -34,30 +35,41 @@ except ImportError:
 
 from . import utils
 
-# import json as _json
+import json as _json
 # import re as _re
 # import sys as _sys
 
 from . import shared
 
+_BASE_URL_ = 'https://query2.finance.yahoo.com'
+_SCRAPE_URL_ = 'https://finance.yahoo.com/quote'
+_ROOT_URL_ = 'https://finance.yahoo.com'
+
 
 class TickerBase():
-    def __init__(self, ticker):
+    def __init__(self, ticker, session=None):
         self.ticker = ticker.upper()
+        self.session = session
         self._history = None
-        self._base_url = 'https://query1.finance.yahoo.com'
-        self._scrape_url = 'https://finance.yahoo.com/quote'
+        self._base_url = _BASE_URL_
+        self._scrape_url = _SCRAPE_URL_
 
         self._fundamentals = False
         self._info = None
+        self._analysis = None
         self._sustainability = None
         self._recommendations = None
         self._major_holders = None
         self._institutional_holders = None
+        self._mutualfund_holders = None
         self._isin = None
+        self._news = []
+        self._shares = None
 
         self._calendar = None
         self._expirations = {}
+        self._earnings_dates = None
+        self._earnings_history = None
 
         self._earnings = {
             "yearly": utils.empty_df(),
@@ -72,10 +84,30 @@ class TickerBase():
             "yearly": utils.empty_df(),
             "quarterly": utils.empty_df()}
 
+        # accept isin as ticker
+        if utils.is_isin(self.ticker):
+            self.ticker = utils.get_ticker_by_isin(self.ticker, None, session)
+
+    def stats(self, proxy=None):
+        # setup proxy in requests format
+        if proxy is not None:
+            if isinstance(proxy, dict) and "https" in proxy:
+                proxy = proxy["https"]
+            proxy = {"https": proxy}
+
+        if self._fundamentals:
+            return
+
+        ticker_url = "{}/{}".format(self._scrape_url, self.ticker)
+
+        # get info and sustainability
+        data = utils.get_json(ticker_url, proxy, self.session)
+        return data
+
     def history(self, period="1mo", interval="1d",
                 start=None, end=None, prepost=False, actions=True,
                 auto_adjust=True, back_adjust=False,
-                proxy=None, rounding=True, tz=None, **kwargs):
+                proxy=None, rounding=False, tz=None, timeout=None, **kwargs):
         """
         :Parameters:
             period : str
@@ -105,6 +137,10 @@ class TickerBase():
             tz: str
                 Optional timezone locale for dates.
                 (default data is returned as non-localized dates)
+            timeout: None or float
+                If not None stops waiting for a response after given number of
+                seconds. (Can also be a fraction of a second e.g. 0.01)
+                Default is None.
             **kwargs: dict
                 debug: bool
                     Optional. If passed as False, will suppress
@@ -112,20 +148,22 @@ class TickerBase():
         """
 
         if start or period is None or period.lower() == "max":
-            if start is None:
-                start = -2208988800
-            elif isinstance(start, _datetime.datetime):
-                start = int(_time.mktime(start.timetuple()))
-            else:
-                start = int(_time.mktime(
-                    _time.strptime(str(start), '%Y-%m-%d')))
             if end is None:
                 end = int(_time.time())
             elif isinstance(end, _datetime.datetime):
                 end = int(_time.mktime(end.timetuple()))
             else:
-                end = int(_time.mktime(_time.strptime(str(end), '%Y-%m-%d')))
-
+                end = int(_time.mktime(_time.strptime(str(end), '%Y-%m-%d'))) 
+            if start is None:
+                if interval=="1m":
+                    start = end - 604800 # Subtract 7 days 
+                else:
+                    start = -631159200
+            elif isinstance(start, _datetime.datetime):
+                start = int(_time.mktime(start.timetuple()))
+            else:
+                start = int(_time.mktime(
+                    _time.strptime(str(start), '%Y-%m-%d')))
             params = {"period1": start, "period2": end}
         else:
             period = period.lower()
@@ -147,12 +185,27 @@ class TickerBase():
 
         # Getting data from json
         url = "{}/v8/finance/chart/{}".format(self._base_url, self.ticker)
-        data = _requests.get(url=url, params=params, proxies=proxy)
-        if "Will be right back" in data.text:
-            raise RuntimeError("*** YAHOO! FINANCE IS CURRENTLY DOWN! ***\n"
-                               "Our engineers are working quickly to resolve "
-                               "the issue. Thank you for your patience.")
-        data = data.json()
+
+        session = self.session or _requests
+
+        data = None
+
+        try:
+            data = session.get(
+                url=url,
+                params=params,
+                proxies=proxy,
+                headers=utils.user_agent_headers,
+                timeout=timeout
+            )
+            if "Will be right back" in data.text or data is None:
+                raise RuntimeError("*** YAHOO! FINANCE IS CURRENTLY DOWN! ***\n"
+                                   "Our engineers are working quickly to resolve "
+                                   "the issue. Thank you for your patience.")
+
+            data = data.json()
+        except Exception:
+            pass
 
         # Work with errors
         debug_mode = True
@@ -160,6 +213,14 @@ class TickerBase():
             debug_mode = kwargs["debug"]
 
         err_msg = "No data found for this date range, symbol may be delisted"
+
+        if data is None or not type(data) is dict or 'status_code' in data.keys():
+            shared._DFS[self.ticker] = utils.empty_df()
+            shared._ERRORS[self.ticker] = err_msg
+            if "many" not in kwargs and debug_mode:
+                print('- %s: %s' % (self.ticker, err_msg))
+            return utils.empty_df()
+
         if "chart" in data and data["chart"]["error"]:
             err_msg = data["chart"]["error"]["description"]
             shared._DFS[self.ticker] = utils.empty_df()
@@ -206,10 +267,20 @@ class TickerBase():
             except Exception:
                 pass
 
-        if auto_adjust:
-            quotes = utils.auto_adjust(quotes)
-        elif back_adjust:
-            quotes = utils.back_adjust(quotes)
+        try:
+            if auto_adjust:
+                quotes = utils.auto_adjust(quotes)
+            elif back_adjust:
+                quotes = utils.back_adjust(quotes)
+        except Exception as e:
+            if auto_adjust:
+                err_msg = "auto_adjust failed with %s" % e
+            else:
+                err_msg = "back_adjust failed with %s" % e
+            shared._DFS[self.ticker] = utils.empty_df()
+            shared._ERRORS[self.ticker] = err_msg
+            if "many" not in kwargs and debug_mode:
+                print('- %s: %s' % (self.ticker, err_msg))
 
         if rounding:
             quotes = _np.round(quotes, data[
@@ -232,11 +303,17 @@ class TickerBase():
 
         if params["interval"][-1] == "m":
             df.index.name = "Datetime"
+        elif params["interval"] == "1h":
+            pass
         else:
             df.index = _pd.to_datetime(df.index.date)
             if tz is not None:
                 df.index = df.index.tz_localize(tz)
             df.index.name = "Date"
+
+        # duplicates and missing rows cleanup
+        df.dropna(how='all', inplace=True)
+        df = df[~df.index.duplicated(keep='first')]
 
         self._history = df.copy()
 
@@ -275,47 +352,104 @@ class TickerBase():
         if self._fundamentals:
             return
 
+        ticker_url = "{}/{}".format(self._scrape_url, self.ticker)
+
         # get info and sustainability
-        url = '%s/%s' % (self._scrape_url, self.ticker)
-        data = utils.get_json(url, proxy)
+        data = utils.get_json(ticker_url, proxy, self.session)
 
         # holders
-        url = "{}/{}/holders".format(self._scrape_url, self.ticker)
-        holders = _pd.read_html(url)
-        self._major_holders = holders[0]
-        self._institutional_holders = holders[1]
-        if 'Date Reported' in self._institutional_holders:
-            self._institutional_holders['Date Reported'] = _pd.to_datetime(
-                self._institutional_holders['Date Reported'])
-        if '% Out' in self._institutional_holders:
-            self._institutional_holders['% Out'] = self._institutional_holders[
-                '% Out'].str.replace('%', '').astype(float)/100
+        try:
+            resp = utils.get_html(ticker_url + '/holders', proxy, self.session)
+            holders = _pd.read_html(resp)
+        except Exception:
+            holders = []
+
+        if len(holders) >= 3:
+            self._major_holders = holders[0]
+            self._institutional_holders = holders[1]
+            self._mutualfund_holders = holders[2]
+        elif len(holders) >= 2:
+            self._major_holders = holders[0]
+            self._institutional_holders = holders[1]
+        elif len(holders) >= 1:
+            self._major_holders = holders[0]
+
+        # self._major_holders = holders[0]
+        # self._institutional_holders = holders[1]
+
+        if self._institutional_holders is not None:
+            if 'Date Reported' in self._institutional_holders:
+                self._institutional_holders['Date Reported'] = _pd.to_datetime(
+                    self._institutional_holders['Date Reported'])
+            if '% Out' in self._institutional_holders:
+                self._institutional_holders['% Out'] = self._institutional_holders[
+                    '% Out'].str.replace('%', '').astype(float) / 100
+
+        if self._mutualfund_holders is not None:
+            if 'Date Reported' in self._mutualfund_holders:
+                self._mutualfund_holders['Date Reported'] = _pd.to_datetime(
+                    self._mutualfund_holders['Date Reported'])
+            if '% Out' in self._mutualfund_holders:
+                self._mutualfund_holders['% Out'] = self._mutualfund_holders[
+                    '% Out'].str.replace('%', '').astype(float) / 100
 
         # sustainability
         d = {}
-        if isinstance(data.get('esgScores'), dict):
-            for item in data['esgScores']:
-                if not isinstance(data['esgScores'][item], (dict, list)):
-                    d[item] = data['esgScores'][item]
+        try:
+            if isinstance(data.get('esgScores'), dict):
+                for item in data['esgScores']:
+                    if not isinstance(data['esgScores'][item], (dict, list)):
+                        d[item] = data['esgScores'][item]
 
-            s = _pd.DataFrame(index=[0], data=d)[-1:].T
-            s.columns = ['Value']
-            s.index.name = '%.f-%.f' % (
-                s[s.index == 'ratingYear']['Value'].values[0],
-                s[s.index == 'ratingMonth']['Value'].values[0])
+                s = _pd.DataFrame(index=[0], data=d)[-1:].T
+                s.columns = ['Value']
+                s.index.name = '%.f-%.f' % (
+                    s[s.index == 'ratingYear']['Value'].values[0],
+                    s[s.index == 'ratingMonth']['Value'].values[0])
 
-            self._sustainability = s[~s.index.isin(
-                ['maxAge', 'ratingYear', 'ratingMonth'])]
+                self._sustainability = s[~s.index.isin(
+                    ['maxAge', 'ratingYear', 'ratingMonth'])]
+        except Exception:
+            pass
 
         # info (be nice to python 2)
         self._info = {}
-        items = ['summaryProfile', 'summaryDetail', 'quoteType',
-                 'defaultKeyStatistics', 'assetProfile', 'summaryDetail']
-        for item in items:
-            if isinstance(data.get(item), dict):
-                self._info.update(data[item])
+        try:
+            items = ['summaryProfile', 'financialData', 'quoteType',
+                     'defaultKeyStatistics', 'assetProfile', 'summaryDetail']
+            for item in items:
+                if isinstance(data.get(item), dict):
+                    self._info.update(data[item])
+        except Exception:
+            pass
 
-        self._info['regularMarketPrice'] = self._info['regularMarketOpen']
+       # For ETFs, provide this valuable data: the top holdings of the ETF
+        try:
+            if 'topHoldings' in data:
+                self._info.update(data['topHoldings'])
+        except Exception:
+            pass
+
+        try:
+            if not isinstance(data.get('summaryDetail'), dict):
+                # For some reason summaryDetail did not give any results. The price dict usually has most of the same info
+                self._info.update(data.get('price', {}))
+        except Exception:
+            pass
+
+        try:
+            # self._info['regularMarketPrice'] = self._info['regularMarketOpen']
+            self._info['regularMarketPrice'] = data.get('price', {}).get(
+                'regularMarketPrice', self._info.get('regularMarketOpen', None))
+        except Exception:
+            pass
+
+        try:
+            self._info['preMarketPrice'] = data.get('price', {}).get(
+                'preMarketPrice', self._info.get('preMarketPrice', None))
+        except Exception:
+            pass
+
         self._info['logo_url'] = ""
         try:
             domain = self._info['website'].split(
@@ -351,7 +485,7 @@ class TickerBase():
             pass
 
         # get fundamentals
-        data = utils.get_json(url+'/financials', proxy)
+        data = utils.get_json(ticker_url + '/financials', proxy, self.session)
 
         # generic patterns
         for key in (
@@ -359,88 +493,199 @@ class TickerBase():
             (self._balancesheet, 'balanceSheet', 'balanceSheetStatements'),
             (self._financials, 'incomeStatement', 'incomeStatementHistory')
         ):
-
             item = key[1] + 'History'
             if isinstance(data.get(item), dict):
-                key[0]['yearly'] = cleanup(data[item][key[2]])
+                try:
+                    key[0]['yearly'] = cleanup(data[item][key[2]])
+                except Exception:
+                    pass
 
-            item = key[1]+'HistoryQuarterly'
+            item = key[1] + 'HistoryQuarterly'
             if isinstance(data.get(item), dict):
-                key[0]['quarterly'] = cleanup(data[item][key[2]])
+                try:
+                    key[0]['quarterly'] = cleanup(data[item][key[2]])
+                except Exception:
+                    pass
 
         # earnings
         if isinstance(data.get('earnings'), dict):
-            earnings = data['earnings']['financialsChart']
-            df = _pd.DataFrame(earnings['yearly']).set_index('date')
-            df.columns = utils.camel2title(df.columns)
-            df.index.name = 'Year'
-            self._earnings['yearly'] = df
+            try:
+                earnings = data['earnings']['financialsChart']
+                earnings['financialCurrency'] = 'USD' if 'financialCurrency' not in data['earnings'] else data['earnings']['financialCurrency']
+                self._earnings['financialCurrency'] = earnings['financialCurrency']
+                df = _pd.DataFrame(earnings['yearly']).set_index('date')
+                df.columns = utils.camel2title(df.columns)
+                df.index.name = 'Year'
+                self._earnings['yearly'] = df
 
-            df = _pd.DataFrame(earnings['quarterly']).set_index('date')
-            df.columns = utils.camel2title(df.columns)
-            df.index.name = 'Quarter'
-            self._earnings['quarterly'] = df
+                df = _pd.DataFrame(earnings['quarterly']).set_index('date')
+                df.columns = utils.camel2title(df.columns)
+                df.index.name = 'Quarter'
+                self._earnings['quarterly'] = df
+            except Exception:
+                pass
+
+        # shares outstanding
+        try:
+            shares = _pd.DataFrame(data['annualBasicAverageShares'])
+            shares['Year'] = shares['asOfDate'].agg(lambda x: int(x[:4]))
+            shares.set_index('Year', inplace=True)
+            shares.drop(columns=['dataId', 'asOfDate',
+                        'periodType', 'currencyCode'], inplace=True)
+            shares.rename(
+                columns={'reportedValue': "BasicShares"}, inplace=True)
+            self._shares = shares
+        except Exception:
+            pass
+
+        # Analysis
+        data = utils.get_json(ticker_url + '/analysis', proxy, self.session)
+
+        if isinstance(data.get('earningsTrend'), dict):
+            try:
+                analysis = _pd.DataFrame(data['earningsTrend']['trend'])
+                analysis['endDate'] = _pd.to_datetime(analysis['endDate'])
+                analysis.set_index('period', inplace=True)
+                analysis.index = analysis.index.str.upper()
+                analysis.index.name = 'Period'
+                analysis.columns = utils.camel2title(analysis.columns)
+
+                dict_cols = []
+
+                for idx, row in analysis.iterrows():
+                    for colname, colval in row.items():
+                        if isinstance(colval, dict):
+                            dict_cols.append(colname)
+                            for k, v in colval.items():
+                                new_colname = colname + ' ' + \
+                                    utils.camel2title([k])[0]
+                                analysis.loc[idx, new_colname] = v
+
+                self._analysis = analysis[[
+                    c for c in analysis.columns if c not in dict_cols]]
+            except Exception:
+                pass
+
+        # Complementary key-statistics (currently fetching the important trailingPegRatio which is the value shown in the website)
+        res = {}
+        try:
+            my_headers = {'user-agent': 'curl/7.55.1', 'accept': 'application/json', 'content-type': 'application/json',
+                          'referer': 'https://finance.yahoo.com/', 'cache-control': 'no-cache', 'connection': 'close'}
+            p = _re.compile(r'root\.App\.main = (.*);')
+            r = _requests.session().get('https://finance.yahoo.com/quote/{}/key-statistics?p={}'.format(self.ticker,
+                                                                                                        self.ticker), headers=my_headers)
+            q_results = {}
+            my_qs_keys = ['pegRatio']  # QuoteSummaryStore
+            # , 'quarterlyPegRatio']  # QuoteTimeSeriesStore
+            my_ts_keys = ['trailingPegRatio']
+
+            # Complementary key-statistics
+            data = _json.loads(p.findall(r.text)[0])
+            key_stats = data['context']['dispatcher']['stores']['QuoteTimeSeriesStore']
+            q_results.setdefault(self.ticker, [])
+            for i in my_ts_keys:
+                # j=0
+                try:
+                    # res = {i: key_stats['timeSeries'][i][1]['reportedValue']['raw']}
+                    # We need to loop over multiple items, if they exist: 0,1,2,..
+                    zzz = key_stats['timeSeries'][i]
+                    for j in range(len(zzz)):
+                        if key_stats['timeSeries'][i][j]:
+                            res = {i: key_stats['timeSeries']
+                                   [i][j]['reportedValue']['raw']}
+                            q_results[self.ticker].append(res)
+
+                # print(res)
+                # q_results[ticker].append(res)
+                except:
+                    q_results[ticker].append({i: np.nan})
+
+            res = {'Company': ticker}
+            q_results[ticker].append(res)
+        except Exception:
+            pass
+
+        if 'trailingPegRatio' in res:
+            self._info['trailingPegRatio'] = res['trailingPegRatio']
 
         self._fundamentals = True
 
     def get_recommendations(self, proxy=None, as_dict=False, *args, **kwargs):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._recommendations
         if as_dict:
             return data.to_dict()
         return data
 
     def get_calendar(self, proxy=None, as_dict=False, *args, **kwargs):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._calendar
         if as_dict:
             return data.to_dict()
         return data
 
     def get_major_holders(self, proxy=None, as_dict=False, *args, **kwargs):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._major_holders
         if as_dict:
             return data.to_dict()
         return data
 
     def get_institutional_holders(self, proxy=None, as_dict=False, *args, **kwargs):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._institutional_holders
-        if as_dict:
-            return data.to_dict()
-        return data
+        if data is not None:
+            if as_dict:
+                return data.to_dict()
+            return data
+
+    def get_mutualfund_holders(self, proxy=None, as_dict=False, *args, **kwargs):
+        self._get_fundamentals(proxy=proxy)
+        data = self._mutualfund_holders
+        if data is not None:
+            if as_dict:
+                return data.to_dict()
+            return data
 
     def get_info(self, proxy=None, as_dict=False, *args, **kwargs):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._info
         if as_dict:
             return data.to_dict()
         return data
 
     def get_sustainability(self, proxy=None, as_dict=False, *args, **kwargs):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._sustainability
         if as_dict:
             return data.to_dict()
         return data
 
     def get_earnings(self, proxy=None, as_dict=False, freq="yearly"):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._earnings[freq]
+        if as_dict:
+            dict_data = data.to_dict()
+            dict_data['financialCurrency'] = 'USD' if 'financialCurrency' not in self._earnings else self._earnings['financialCurrency']
+            return dict_data
+        return data
+
+    def get_analysis(self, proxy=None, as_dict=False, *args, **kwargs):
+        self._get_fundamentals(proxy=proxy)
+        data = self._analysis
         if as_dict:
             return data.to_dict()
         return data
 
     def get_financials(self, proxy=None, as_dict=False, freq="yearly"):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._financials[freq]
         if as_dict:
             return data.to_dict()
         return data
 
     def get_balancesheet(self, proxy=None, as_dict=False, freq="yearly"):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._balancesheet[freq]
         if as_dict:
             return data.to_dict()
@@ -450,7 +695,7 @@ class TickerBase():
         return self.get_balancesheet(proxy, as_dict, freq)
 
     def get_cashflow(self, proxy=None, as_dict=False, freq="yearly"):
-        self._get_fundamentals(proxy)
+        self._get_fundamentals(proxy=proxy)
         data = self._cashflow[freq]
         if as_dict:
             return data.to_dict()
@@ -459,20 +704,33 @@ class TickerBase():
     def get_dividends(self, proxy=None):
         if self._history is None:
             self.history(period="max", proxy=proxy)
-        dividends = self._history["Dividends"]
-        return dividends[dividends != 0]
+        if self._history is not None and "Dividends" in self._history:
+            dividends = self._history["Dividends"]
+            return dividends[dividends != 0]
+        return []
 
     def get_splits(self, proxy=None):
         if self._history is None:
             self.history(period="max", proxy=proxy)
-        splits = self._history["Stock Splits"]
-        return splits[splits != 0]
+        if self._history is not None and "Stock Splits" in self._history:
+            splits = self._history["Stock Splits"]
+            return splits[splits != 0]
+        return []
 
     def get_actions(self, proxy=None):
         if self._history is None:
             self.history(period="max", proxy=proxy)
-        actions = self._history[["Dividends", "Stock Splits"]]
-        return actions[actions != 0].dropna(how='all').fillna(0)
+        if self._history is not None and "Dividends" in self._history and "Stock Splits" in self._history:
+            actions = self._history[["Dividends", "Stock Splits"]]
+            return actions[actions != 0].dropna(how='all').fillna(0)
+        return []
+
+    def get_shares(self, proxy=None, as_dict=False, *args, **kwargs):
+        self._get_fundamentals(proxy=proxy)
+        data = self._shares
+        if as_dict:
+            return data.to_dict()
+        return data
 
     def get_isin(self, proxy=None):
         # *** experimental ***
@@ -499,7 +757,12 @@ class TickerBase():
         url = 'https://markets.businessinsider.com/ajax/' \
               'SearchController_Suggest?max_results=25&query=%s' \
             % urlencode(q)
-        data = _requests.get(url=url, proxies=proxy).text
+        session = self.session or _requests
+        data = session.get(
+            url=url,
+            proxies=proxy,
+            headers=utils.user_agent_headers
+        ).text
 
         search_str = '"{}|'.format(ticker)
         if search_str not in data:
@@ -514,3 +777,149 @@ class TickerBase():
 
         self._isin = data.split(search_str)[1].split('"')[0].split('|')[0]
         return self._isin
+
+    def get_news(self, proxy=None):
+        if self._news:
+            return self._news
+
+        # setup proxy in requests format
+        if proxy is not None:
+            if isinstance(proxy, dict) and "https" in proxy:
+                proxy = proxy["https"]
+            proxy = {"https": proxy}
+
+        # Getting data from json
+        url = "{}/v1/finance/search?q={}".format(self._base_url, self.ticker)
+        session = self.session or _requests
+        data = session.get(
+            url=url,
+            proxies=proxy,
+            headers=utils.user_agent_headers
+        )
+        if "Will be right back" in data.text:
+            raise RuntimeError("*** YAHOO! FINANCE IS CURRENTLY DOWN! ***\n"
+                               "Our engineers are working quickly to resolve "
+                               "the issue. Thank you for your patience.")
+        data = data.json()
+
+        # parse news
+        self._news = data.get("news", [])
+        return self._news
+
+    def get_earnings_dates(self, proxy=None):
+        if self._earnings_dates is not None:
+            return self._earnings_dates
+
+        # setup proxy in requests format
+        if proxy is not None:
+            if isinstance(proxy, dict) and "https" in proxy:
+                proxy = proxy["https"]
+            proxy = {"https": proxy}
+
+        page_size = 100  # YF caps at 100, don't go higher
+        page_offset = 0
+        dates = None
+        while True:
+            url = "{}/calendar/earnings?symbol={}&offset={}&size={}".format(
+                _ROOT_URL_, self.ticker, page_offset, page_size)
+
+            session = self.session or _requests
+            data = session.get(
+                url=url,
+                proxies=proxy,
+                headers=utils.user_agent_headers
+            ).text
+
+            if "Will be right back" in data:
+                raise RuntimeError("*** YAHOO! FINANCE IS CURRENTLY DOWN! ***\n"
+                                   "Our engineers are working quickly to resolve "
+                                   "the issue. Thank you for your patience.")
+
+            try:
+                data = _pd.read_html(data)[0]
+            except ValueError:
+                if page_offset == 0:
+                    # Should not fail on first page
+                    if "Showing Earnings for:" in data:
+                        # Actually YF was successful, problem is company doesn't have earnings history
+                        dates = utils.empty_earnings_dates_df()
+                break
+
+            if dates is None:
+                dates = data
+            else:
+                dates = _pd.concat([dates, data], axis=0)
+            page_offset += page_size
+
+        if dates is None:
+            raise Exception("No data found, symbol may be delisted")
+        dates = dates.reset_index(drop=True)
+
+        # Drop redundant columns
+        dates = dates.drop(["Symbol", "Company"], axis=1)
+
+        # Convert types
+        for cn in ["EPS Estimate", "Reported EPS", "Surprise(%)"]:
+            dates.loc[dates[cn] == '-', cn] = "NaN"
+            dates[cn] = dates[cn].astype(float)
+
+        # Convert % to range 0->1:
+        dates["Surprise(%)"] *= 0.01
+
+        # Parse earnings date string
+        cn = "Earnings Date"
+        # - remove AM/PM and timezone from date string
+        tzinfo = dates[cn].str.extract('([AP]M[a-zA-Z]*)$')
+        dates[cn] = dates[cn].replace(' [AP]M[a-zA-Z]*$', '', regex=True)
+        # - split AM/PM from timezone
+        tzinfo = tzinfo[0].str.extract('([AP]M)([a-zA-Z]*)', expand=True)
+        tzinfo.columns = ["AM/PM", "TZ"]
+        # - combine and parse
+        dates[cn] = dates[cn] + ' ' + tzinfo["AM/PM"]
+        dates[cn] = _pd.to_datetime(dates[cn], format="%b %d, %Y, %I %p")
+        # - instead of attempting decoding of ambiguous timezone abbreviation, just use 'info':
+        dates[cn] = dates[cn].dt.tz_localize(
+            tz=self.info["exchangeTimezoneName"])
+
+        dates = dates.set_index("Earnings Date")
+
+        self._earnings_dates = dates
+
+        return dates
+
+    def get_earnings_history(self, proxy=None):
+        if self._earnings_history:
+            return self._earnings_history
+
+        # setup proxy in requests format
+        if proxy is not None:
+            if isinstance(proxy, dict) and "https" in proxy:
+                proxy = proxy["https"]
+            proxy = {"https": proxy}
+
+        url = "{}/calendar/earnings?symbol={}".format(_ROOT_URL_, self.ticker)
+        session = self.session or _requests
+        data = session.get(
+            url=url,
+            proxies=proxy,
+            headers=utils.user_agent_headers
+        ).text
+
+        if "Will be right back" in data:
+            raise RuntimeError("*** YAHOO! FINANCE IS CURRENTLY DOWN! ***\n"
+                               "Our engineers are working quickly to resolve "
+                               "the issue. Thank you for your patience.")
+
+        try:
+            # read_html returns a list of pandas Dataframes of all the tables in `data`
+            data = _pd.read_html(data)[0]
+            data.replace("-", _np.nan, inplace=True)
+
+            data['EPS Estimate'] = _pd.to_numeric(data['EPS Estimate'])
+            data['Reported EPS'] = _pd.to_numeric(data['Reported EPS'])
+            self._earnings_history = data
+        # if no tables are found a ValueError is thrown
+        except ValueError:
+            print("Could not find data for {}.".format(self.ticker))
+            return
+        return data