+++ yfinance/multi.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-# Yahoo! Finance market data downloader (+fix for Pandas Datareader)
+# yfinance - market data downloader
 # https://github.com/ranaroussi/yfinance
 #
 # Copyright 2017-2019 Ran Aroussi
@@ -31,8 +31,8 @@ from . import shared
 
 def download(tickers, start=None, end=None, actions=False, threads=True,
              group_by='column', auto_adjust=False, back_adjust=False,
-             progress=True, period="max", interval="1d", prepost=False,
-             proxy=None, rounding=False, **kwargs):
+             progress=True, period="max", show_errors=True, interval="1d", prepost=False,
+             proxy=None, rounding=False, timeout=None, **kwargs):
     """Download yahoo tickers
     :Parameters:
         tickers : str, list
@@ -64,11 +64,28 @@ def download(tickers, start=None, end=None, actions=False, threads=True,
             Optional. Proxy server URL scheme. Default is None
         rounding: bool
             Optional. Round values to 2 decimal places?
+        show_errors: bool
+            Optional. Doesn't print errors if True
+        timeout: None or float
+            If not None stops waiting for a response after given number of
+            seconds. (Can also be a fraction of a second e.g. 0.01)
     """
 
     # create ticker list
     tickers = tickers if isinstance(
         tickers, (list, set, tuple)) else tickers.replace(',', ' ').split()
+
+    # accept isin as ticker
+    shared._ISINS = {}
+    _tickers_ = []
+    for ticker in tickers:
+        if utils.is_isin(ticker):
+            isin = ticker
+            ticker = utils.get_ticker_by_isin(ticker, proxy)
+            shared._ISINS[ticker] = isin
+        _tickers_.append(ticker)
+
+    tickers = _tickers_
 
     tickers = list(set([ticker.upper() for ticker in tickers]))
 
@@ -90,7 +107,7 @@ def download(tickers, start=None, end=None, actions=False, threads=True,
                                    actions=actions, auto_adjust=auto_adjust,
                                    back_adjust=back_adjust,
                                    progress=(progress and i > 0), proxy=proxy,
-                                   rounding=rounding)
+                                   rounding=rounding, timeout=timeout)
         while len(shared._DFS) < len(tickers):
             _time.sleep(0.01)
 
@@ -100,7 +117,8 @@ def download(tickers, start=None, end=None, actions=False, threads=True,
             data = _download_one(ticker, period=period, interval=interval,
                                  start=start, end=end, prepost=prepost,
                                  actions=actions, auto_adjust=auto_adjust,
-                                 back_adjust=back_adjust, rounding=rounding)
+                                 back_adjust=back_adjust, proxy=proxy,
+                                 rounding=rounding, timeout=timeout)
             shared._DFS[ticker.upper()] = data
             if progress:
                 shared._PROGRESS_BAR.animate()
@@ -108,7 +126,7 @@ def download(tickers, start=None, end=None, actions=False, threads=True,
     if progress:
         shared._PROGRESS_BAR.completed()
 
-    if shared._ERRORS:
+    if shared._ERRORS and show_errors:
         print('\n%.f Failed download%s:' % (
             len(shared._ERRORS), 's' if len(shared._ERRORS) > 1 else ''))
         # print(shared._ERRORS)
@@ -116,7 +134,8 @@ def download(tickers, start=None, end=None, actions=False, threads=True,
                          v for v in list(shared._ERRORS.items())]))
 
     if len(tickers) == 1:
-        return shared._DFS[tickers[0]]
+        ticker = tickers[0]
+        return shared._DFS[shared._ISINS.get(ticker, ticker)]
 
     try:
         data = _pd.concat(shared._DFS.values(), axis=1,
@@ -125,6 +144,9 @@ def download(tickers, start=None, end=None, actions=False, threads=True,
         _realign_dfs()
         data = _pd.concat(shared._DFS.values(), axis=1,
                           keys=shared._DFS.keys())
+
+    # switch names back to isins if applicable
+    data.rename(columns=shared._ISINS, inplace=True)
 
     if group_by == 'column':
         data.columns = data.columns.swaplevel(0, 1)
@@ -161,10 +183,11 @@ def _download_one_threaded(ticker, start=None, end=None,
                            auto_adjust=False, back_adjust=False,
                            actions=False, progress=True, period="max",
                            interval="1d", prepost=False, proxy=None,
-                           rounding=False):
+                           rounding=False, timeout=None):
 
     data = _download_one(ticker, start, end, auto_adjust, back_adjust,
-                         actions, period, interval, prepost, proxy, rounding)
+                         actions, period, interval, prepost, proxy, rounding,
+                         timeout)
     shared._DFS[ticker.upper()] = data
     if progress:
         shared._PROGRESS_BAR.animate()
@@ -173,10 +196,12 @@ def _download_one_threaded(ticker, start=None, end=None,
 def _download_one(ticker, start=None, end=None,
                   auto_adjust=False, back_adjust=False,
                   actions=False, period="max", interval="1d",
-                  prepost=False, proxy=None, rounding=False):
+                  prepost=False, proxy=None, rounding=False,
+                  timeout=None):
 
     return Ticker(ticker).history(period=period, interval=interval,
                                   start=start, end=end, prepost=prepost,
                                   actions=actions, auto_adjust=auto_adjust,
                                   back_adjust=back_adjust, proxy=proxy,
-                                  rounding=rounding, many=True)
+                                  rounding=rounding, many=True,
+                                  timeout=timeout)